+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## `v1.0.2` (contracts deployed 2021-04-02; all code pushed 2021-04-17)
+
+* Created `Keep3rV2PriceOracle`.
+
 ## `v1.0.1` (contracts deployed 2021-03-21; all code pushed 2021-04-17)
 
 * Added `_callPool` functions to `FuseFeeDistributor`.